requirements.txt
ultralytics
streamlit
numpy
Pillow
torch
torchvision
